get_envirem_clim: Get bioclimatic data from envirem

Get bioclimatic layers from ENVIREM for present and past scenarios

get_envirem_clim(
  period,
  gcm = NULL,
  region,
  resolution,
  fmt,
  sv_dir = getwd(),
  load2r
)

Arguments

period
Period. Possible values are: "current" ( 1960 - 1990),"holo" (Mid-Holocene ~ 6000 years ago),"lgm" (Last Glacial Maximum ~ 22000 years ago).

gcm
Global circulation model. Posible values: NULL, "ccsm4", "miroc_esm", and, "mpi_esm_p".

region
Geographic region. Posibles values are: "Africa", "Australia", "Eurasia", "Europe", "global", "NAmerica", "NewWorld", "Pacific", and, "SAmerica".

resolution
Resolution of layers. Posible values are "10arcmin","5arcmin","2.5arcmin", and, "30arcsec".

fmt
File format. Possible values are "bil" and "geotiff".

sv_dir
Path to the directory where the layers will be saved. The default is the working directory of the R session.

load2r
Logical. Load layers into R?

Details

For more details visit https://envirem.github.io/

References

Title P.O., Bemmels J.B. 2018. ENVIREM: an expanded set of bioclimatic and topographic variables increases flexibility and improves the performance of ecological niche modeling. Ecography. 41:291-307.

See also

Examples

if (FALSE) {
# For a compleate list of posible values of the function
# for current period
# clim_data_current <- base::readRDS(file.path(system.file("extdata",
#                                                   package = "ntbox"),
#                                                  "envirem_clim_current.rds"))
# For a compleate list of posible values of the function
# for pass perids (holo,lgm)

clim_eu_lgm_ccsm4_10min <- get_envirem_clim(period="lgm",
                                            gcm="ccsm4",
                                            region="Europe",
                                            resolution= "10arcmin",
                                            fmt ="geotiff",
                                            sv_dir = "~/Desktop",
                                            load2r=TRUE)
}